Add <c-t> mapping from rsi.vim
Tried rsi.vim, but it introduces undesired mappings in insert mode
(which do break important use cases) and doesn't provide useful
mappings for command-line mode.

emacscommandline was missing this bit and is otherwise OK (except
that when debugging it should be used carefully or function calls
of mappings get debugged too).